Протокол телеметрії v2: труба 101, 16-бітні id, розкладка блоками
Борт FIBER_WIND замінив стару пару плат, і разом з ним прийшла нова
розкладка телеметрії. Змінилось три речі, і кожна ламає сумісність.

ТРУБА. Перший байт кадру тепер pipe_id у глобальному просторі: 1..100
керування, 101..254 телеметрія, 255 хартбіт. Станція слухає трубу 101 і
фільтрує саме за нею, а не за портом: порт каже, куди прийшло, труба — що
це насправді. Саме кадрування не змінилось, CRC звірено з трьома
контрольними значеннями специфікації.

ЗАПИС. Було [id:1][float32], стало [id:2 LE][float32 LE]. Id виріс не про
запас: вживані номери сягають 526, у байт це не лізе. Сховище відповідно
стало 16-бітним, ємність 1024 слоти.

РОЗКЛАДКА. Блоками по 64 за змістом, із запасом у кінці кожного: нове поле
лягає у свій блок і НІКОЛИ не зсуває сусідів, тож нумерацію більше не
доводиться заморожувати. Перенумеровано 33 канали в основній розкладці OSD
і 32 в другій, жодного без відповідника.

Власні канали станції переїхали з 200..214 у 528..542: старе місце тепер
всередині блоку каналів пультів борта (192..223), і збіг був би тихим —
стан запису читався б як положення стіка.

Канал перемикача пресетів: 65 -> 206, бо RC1_CH15 це 192+14.

Плату POINT злито з польотним контролером, тож її окреме живлення прибрано
з розкладки — тепер це VBAT борта.

Таблиця імен ТЕПЕР ГЕНЕРУЄТЬСЯ з переліку id. Написана руками, вона рано
чи пізно розходиться з ним, а таблиця імен, що розійшлася, гірша за її
відсутність: вона впевнено називає чуже поле своїм.

Перевірено розбором справжнього кадру: id понад байт (323) і простір
станції (528) читаються правильно, CRC-збоїв нуль.

+++ osd/telemetry/vt_telemetry_fetch.h
@@ -25,7 +25,7 @@ inline bool vt_fetch_channel_value(VtTelemetryStorage& storage, int channel_id,
                                     bool is_age_meta, bool is_rate_meta,
                                     float* out_value) {
     if (channel_id < 0) return false;
-    const uint8_t id = static_cast<uint8_t>(channel_id);
+    const uint16_t id = static_cast<uint16_t>(channel_id);
 
     uint32_t age_ms = 0;
     float value = 0.f;
